Found a few more sneaky extension methods trying to hide
